Fix converting rgba from string.

+++ src/Drivers/Color.php
@@ -91,7 +91,7 @@ abstract class Color
             $result[0] = ($matches[1] >= 0 && $matches[1] <= 255) ? (int) ($matches[1]) : 0;
             $result[1] = ($matches[2] >= 0 && $matches[2] <= 255) ? (int) ($matches[2]) : 0;
             $result[2] = ($matches[3] >= 0 && $matches[3] <= 255) ? (int) ($matches[3]) : 0;
-            $result[3] = ($matches[4] >= 0 && $matches[4] <= 1) ? (int) ($matches[4]) : 0;
+            $result[3] = ($matches[4] >= 0 && $matches[4] <= 1) ? (float) ($matches[4]) : 0;
         } else {
             throw InvalidColor::make($colorValue);
         }
